Use Firefox Containers for awssso login workflow
Need to use containers when using multiple AWS SSO Instances
so that the browser doesn't re-use its existing SSO session.
Otherwise, switching SSO Instance (--sso) ends up being broken
because you'll use the current session all the time.

Refs: #374
